LayoutTests/media/video-controls-zoomed-expected.txt

Test controls on zoomed video.

This test only runs in DRT!

EXPECTED (video.controls != 'null') OK
EVENT(load)
EXPECTED (video.paused == 'false') OK
END OF TEST
